Split renderers.py into per-model-family modules
Refactors the large renderers.py (~2350 lines) into a package with separate
modules for each model family:

- base.py: Types, Renderer ABC, utilities
- role_colon.py: RoleColonRenderer
- llama3.py: Llama3Renderer
- qwen3.py: Qwen3*, Qwen3VL* renderers
- deepseek_v3.py: DeepSeekV3* renderers
- kimi_k2.py: KimiK2Renderer
- gpt_oss.py: GptOssRenderer

Backwards compatibility is preserved via __init__.py re-exports.
All imports from tinker_cookbook.renderers continue to work unchanged.
