--- tinker_cookbook/renderers.py
@@ -1,2354 +1,0 @@
-"""
-Use viz_sft_dataset to visualize the output of different renderers. E.g.,
-    python -m tinker_cookbook.supervised.viz_sft_dataset dataset_path=Tulu3Builder renderer_name=role_colon
-"""
-
-import io
-import json
-import logging
-import re
-import urllib.request
-from datetime import datetime
-from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from enum import StrEnum
-from typing import Literal, NotRequired, Optional, Protocol, TypedDict, cast
-
-import pydantic
-import tinker
-import torch
-from PIL import Image
-
-from tinker_cookbook.image_processing_utils import ImageProcessor
-from tinker_cookbook.tokenizer_utils import Tokenizer
-
-logger = logging.getLogger(__name__)
-
-# Tool types are based on kosong (https://github.com/MoonshotAI/kosong).
-
-
-class StrictBase(pydantic.BaseModel):
-    """
-    Pydantic base class that's immutable and doesn't silently ignore extra fields.
-    """
-
-    model_config = pydantic.ConfigDict(frozen=True, extra="forbid")
-
-    def __str__(self) -> str:
-        return repr(self)
-
-
-class ToolCall(StrictBase):
-    """
-    Structured tool invocation following OpenAI/kosong format.
-
-    This represents a request to invoke a tool/function. The structure follows
-    the OpenAI function calling format for compatibility with various LLM APIs.
-
-    Example:
-        tool_call = ToolCall(
-            function=ToolCall.FunctionBody(
-                name="search",
-                arguments='{"query_list": ["python async", "pydantic validation"]}'
-            ),
-            id="call_abc123"
-        )
-    """
-
-    class FunctionBody(pydantic.BaseModel):
-        """
-        Tool call function body containing the tool name and arguments.
-
-        The arguments field must be a valid JSON string that will be parsed
-        by the tool implementation.
-        """
-
-        name: str
-        """The name of the tool to be called."""
-        arguments: str
-        """Arguments of the tool call in JSON string format."""
-
-    type: Literal["function"] = "function"
-    """Tool call type, must be 'function' for compatibility."""
-
-    id: str | None = None
-    """Optional unique identifier for tracking this specific tool call."""
-
-    function: FunctionBody
-    """The function body containing tool name and arguments."""
-
-
-class UnparsedToolCall(StrictBase):
-    """
-    Represents a tool call that failed to parse from model output.
-
-    When a model generates text that looks like a tool call but cannot be
-    parsed (e.g., invalid JSON), this class captures the raw text and error
-    for debugging and optional re-rendering.
-
-    Example:
-        unparsed = UnparsedToolCall(
-            raw_text='<tool_call>{"name": "search", invalid json}</tool_call>',
-            error="Invalid JSON: Expecting property name"
-        )
-    """
-
-    raw_text: str
-    """The original text from the model that failed to parse."""
-
-    error: str
-    """Description of what went wrong during parsing."""
-
-
-class TextPart(TypedDict):
-    """A chunk of text content in a message, usually meant to be visible to the user
-    (unlike ThinkingPart, which is internal reasoning)."""
-
-    type: Literal["text"]
-    text: str
-
-
-class ImagePart(TypedDict):
-    """
-    A chunk of image content in a message.
-    """
-
-    type: Literal["image"]
-    image: str | Image.Image
-
-
-class ThinkingPart(TypedDict):
-    """Model's internal reasoning (chain-of-thought) as a content part."""
-
-    type: Literal["thinking"]
-    thinking: str  # The thinking/reasoning content
-
-
-class ToolCallPart(TypedDict):
-    """Tool/function call as a content part, preserving position in content list."""
-
-    type: Literal["tool_call"]
-    tool_call: ToolCall  # The parsed tool call object
-
-
-class UnparsedToolCallPart(TypedDict):
-    """Tool call that failed to parse, preserving raw text for debugging."""
-
-    type: Literal["unparsed_tool_call"]
-    raw_text: str  # Raw text of the tool call block including tags
-    error: str  # Description of what went wrong during parsing
-
-
-# Container for a part of a multimodal message content
-ContentPart = TextPart | ImagePart | ThinkingPart | ToolCallPart | UnparsedToolCallPart
-
-
-# NOTE: we use a broad type definition for the role to be flexible
-# Common roles are "user", "assistant", "system", "tool"
-Role = str
-
-# Content is a string or a list of parts
-Content = str | list[ContentPart]
-
-
-class Message(TypedDict):
-    """
-    Container for a single turn in a multi-turn conversation.
-
-    Args:
-
-    role: Role
-        String that denotes the source of the message, typically system, user, assistant, and tool.
-    content: Content
-        Content of the message, can be a string, or a list of ContentPart.
-        When content is a list, it can contain TextPart, ImagePart, and ThinkingPart elements.
-        ThinkingPart represents the model's internal reasoning (chain-of-thought).
-    tool_calls: NotRequired[list[ToolCall]]
-        Optional sequence of successfully parsed tool calls generated by the model.
-    unparsed_tool_calls: NotRequired[list[UnparsedToolCall]]
-        Optional sequence of tool calls that failed to parse (e.g., invalid JSON).
-        The raw text is preserved for debugging or re-rendering.
-    trainable: NotRequired[bool]
-        Optional indicator whether this message should contribute to the training loss.
-
-    """
-
-    role: Role
-    content: Content
-
-    tool_calls: NotRequired[list[ToolCall]]
-    unparsed_tool_calls: NotRequired[list["UnparsedToolCall"]]
-    trainable: NotRequired[bool]
-    tool_call_id: NotRequired[str]
-    name: NotRequired[str]
-
-
-@dataclass
-class RenderContext:
-    """
-    Context passed to render_message for rendering a single message.
-
-    This allows renderers to access information about the message's position
-    in the conversation without changing the render_message signature for
-    each new piece of context needed.
-    """
-
-    idx: int
-    """Index of the message in the conversation (0-based)."""
-
-    is_last: bool
-    """Whether this is the last message in the conversation."""
-
-    prev_message: Message | None = None
-    """The previous message in the conversation, if any."""
-
-
-class ToolSpec(TypedDict):
-    """
-    Tool specification following the OpenAI function calling format.
-
-    This represents a tool that can be called by the model, including its name,
-    description, and parameter schema.
-
-    Example:
-        tool_spec: ToolSpec = {
-            "name": "get_weather",
-            "description": "Get the current weather for a location",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "location": {"type": "string", "description": "City name"},
-                },
-                "required": ["location"],
-            },
-        }
-    """
-
-    name: str
-    """The name of the tool."""
-    description: str
-    """A description of what the tool does."""
-    parameters: dict
-    """JSON Schema object describing the tool's parameters."""
-
-
-def ensure_text(content: Content) -> str:
-    """
-    Assert that content is text-only and return it as a string.
-
-    Raises ValueError if content contains images or multiple parts.
-    Use this to validate that message content is text-only before
-    processing it in code paths that don't support multimodal content.
-    """
-    if isinstance(content, str):
-        return content
-    if len(content) == 1 and content[0]["type"] == "text":
-        return content[0]["text"]
-    raise ValueError(f"Expected text content, got multimodal content with {len(content)} parts")
-
-
-def ensure_list(content: Content) -> list[ContentPart]:
-    """Normalize content to list form. Wraps string content in a TextPart."""
-    if isinstance(content, str):
-        return [TextPart(type="text", text=content)]
-    return content
-
-
-def remove_thinking(parts: list[ContentPart]) -> list[ContentPart]:
-    """Filter out ThinkingPart elements from a content part list."""
-    return [p for p in parts if p["type"] != "thinking"]
-
-
-def get_text_content(message: Message) -> str:
-    """Extract text content from message, stripping thinking parts.
-
-    Use this after parse_response when you only need the text output,
-    ignoring any thinking/reasoning content.
-    """
-    content = message["content"]
-    if isinstance(content, str):
-        return content
-    return "".join(p["text"] for p in content if p["type"] == "text")
-
-
-def format_content_as_string(content: Content, separator: str = "\n") -> str:
-    """Format message content as a string, preserving all part types.
-
-    Unlike get_text_content which only extracts text parts, this formats
-    all content parts (thinking, text, tool_call, etc.) as a readable string.
-
-    This is useful for compatibility with APIs that expect string content
-    (e.g., OpenAI Chat Completions API), but we don't recommend it if you
-    need to ensure correctness - prefer working with structured content directly
-    and using build_generation_prompt to convert to tokens.
-
-    Args:
-        content: Message content (string or list of ContentPart).
-        separator: String to join parts with. Default is newline.
-
-    Returns:
-        Formatted string representation of all content parts.
-    """
-    if isinstance(content, str):
-        return content
-
-    parts = []
-    for p in content:
-        if p["type"] == "thinking":
-            parts.append(f"<thinking>{p['thinking']}</thinking>")
-        elif p["type"] == "text":
-            parts.append(p["text"])
-        elif p["type"] == "tool_call":
-            tc = p["tool_call"]
-            parts.append(f"<tool_call>{tc.function.name}({tc.function.arguments})</tool_call>")
-        elif p["type"] == "unparsed_tool_call":
-            parts.append(f"<unparsed_tool_call>{p['raw_text']}</unparsed_tool_call>")
-        else:
-            raise ValueError(f"Unknown content part type: {p['type']}")
-    return separator.join(parts)
-
-
-def _parse_tool_call_json(tool_call_str: str, raw_text: str) -> ToolCall | UnparsedToolCall:
-    """Parse tool call JSON. Returns UnparsedToolCall on failure."""
-    try:
-        tool_call = json.loads(tool_call_str.strip())
-    except json.JSONDecodeError as e:
-        return UnparsedToolCall(raw_text=raw_text, error=f"Invalid JSON: {e}")
-
-    if not isinstance(tool_call, dict):
-        return UnparsedToolCall(raw_text=raw_text, error="Tool call is not a JSON object")
-
-    name = tool_call.get("name")
-    arguments = tool_call.get("arguments")
-    tool_id = tool_call.get("id")
-
-    if not isinstance(name, str):
-        return UnparsedToolCall(raw_text=raw_text, error="Missing or invalid 'name' field")
-    if not isinstance(arguments, dict):
-        return UnparsedToolCall(raw_text=raw_text, error="Missing or invalid 'arguments' field")
-
-    if tool_id is not None and not isinstance(tool_id, str):
-        tool_id = None
-
-    # TODO: arguments is already a dict from json.loads above, but ToolCall.FunctionBody.arguments
-    # expects a JSON string. This round-trip (loads then dumps) is wasteful. Consider changing
-    # FunctionBody.arguments to accept dict directly, or parse tool calls more lazily.
-    # We may want to revisit the decision to store arguments as unparsed JSON string.
-    return ToolCall(
-        function=ToolCall.FunctionBody(name=name, arguments=json.dumps(arguments)),
-        id=tool_id,
-    )
-
-
-def parse_content_blocks(content: str) -> list[ContentPart] | None:
-    """
-    Parse a string with <think>...</think> and <tool_call>...</tool_call> tags.
-
-    Handles interleaved thinking, tool call, and text blocks, returning parts
-    in order. Empty parts are omitted. Failed tool call parses are included as
-    UnparsedToolCallPart to preserve ordering.
-
-    Whitespace is preserved exactly - roundtrip (parse then render) is identity.
-
-    Args:
-        content: String potentially containing <think> and/or <tool_call> blocks.
-
-    Returns:
-        List of ContentPart (ThinkingPart, TextPart, ToolCallPart, UnparsedToolCallPart)
-        in order. Returns None if no special tags are found - caller should use
-        the original string for backward compatibility.
-
-    Example:
-        >>> parse_content_blocks("<think>step 1</think>answer<tool_call>{...}</tool_call>more")
-        [
-            ThinkingPart(type="thinking", thinking="step 1"),
-            TextPart(type="text", text="answer"),
-            ToolCallPart(type="tool_call", tool_call=ToolCall(...)),
-            TextPart(type="text", text="more"),
-        ]
-    """
-    if "<think>" not in content and "<tool_call>" not in content:
-        return None  # No special blocks, caller should use original string
-
-    parts: list[ContentPart] = []
-    pos = 0
-
-    # Pattern to find both <think>...</think> and <tool_call>...</tool_call> blocks
-    pattern = re.compile(r"<think>(.*?)</think>|<tool_call>(.*?)</tool_call>", re.DOTALL)
-
-    for match in pattern.finditer(content):
-        # Add any text before this block (preserve whitespace for identity roundtrip)
-        text_before = content[pos : match.start()]
-        if text_before:  # Skip only truly empty strings
-            parts.append(TextPart(type="text", text=text_before))
-
-        if match.group(1) is not None:
-            # This is a <think> block
-            thinking = match.group(1)
-            if thinking:  # Skip empty thinking blocks
-                parts.append(ThinkingPart(type="thinking", thinking=thinking))
-        else:
-            # This is a <tool_call> block
-            tool_call_json = match.group(2)
-            raw_text = match.group(0)  # Full match including tags
-            parsed = _parse_tool_call_json(tool_call_json, raw_text)
-            if isinstance(parsed, UnparsedToolCall):
-                # Include unparsed tool calls as UnparsedToolCallPart to preserve order
-                parts.append(
-                    UnparsedToolCallPart(
-                        type="unparsed_tool_call",
-                        raw_text=parsed.raw_text,
-                        error=parsed.error,
-                    )
-                )
-            else:
-                parts.append(ToolCallPart(type="tool_call", tool_call=parsed))
-
-        pos = match.end()
-
-    # Add any remaining text after the last block
-    remaining = content[pos:]
-    if remaining:  # Skip only truly empty strings
-        parts.append(TextPart(type="text", text=remaining))
-
-    return parts
-
-
-def parse_think_blocks(content: str) -> list[ContentPart] | None:
-    """
-    Parse a string with only <think>...</think> tags into ThinkingPart/TextPart list.
-
-    This is a simpler version of parse_content_blocks for renderers that use
-    non-standard tool call formats (like DeepSeek's <｜tool▁calls▁begin｜>).
-
-    Whitespace is preserved exactly - roundtrip (parse then render) is identity.
-
-    Args:
-        content: String potentially containing <think>...</think> blocks.
-
-    Returns:
-        List of ThinkingPart and TextPart in order. None if no <think> tags found.
-    """
-    if "<think>" not in content:
-        return None
-
-    parts: list[ContentPart] = []
-    pos = 0
-    pattern = re.compile(r"<think>(.*?)</think>", re.DOTALL)
-
-    for match in pattern.finditer(content):
-        text_before = content[pos : match.start()]
-        if text_before:  # Skip only truly empty strings
-            parts.append(TextPart(type="text", text=text_before))
-
-        thinking = match.group(1)
-        if thinking:  # Skip empty thinking blocks
-            parts.append(ThinkingPart(type="thinking", thinking=thinking))
-
-        pos = match.end()
-
-    remaining = content[pos:]
-    if remaining:  # Skip only truly empty strings
-        parts.append(TextPart(type="text", text=remaining))
-
-    return parts
-
-
-def _tool_call_payload(tool_call: ToolCall) -> dict[str, object]:
-    """Minimal JSON payload for embedding in <tool_call> blocks."""
-    # Convert from nested structure to flat format for compatibility
-    return {
-        "name": tool_call.function.name,
-        "arguments": json.loads(tool_call.function.arguments),
-    }
-
-
-@dataclass(frozen=True)
-class RenderedMessage:
-    """
-    Container for parts of a rendered message, structured for loss masking.
-
-    A rendered message is split into header and output to control which tokens receive
-    training loss. In the simplest case (where the full conversation is formed by
-    concatenation), building a supervised example from messages [m_0, ..., m_{n-1}]
-    produces:
-
-        tokens = BOS + header_0 + output_0 + header_1 + output_1 + ... + header_{n-1} + output_{n-1}
-
-    However, some renderers modify this structure. For example, Qwen3Renderer strips
-    thinking blocks from historical assistant messages. Such renderers must override
-    build_supervised_example to match their build_generation_prompt behavior.
-
-    Attributes:
-        output: What the model generates for this turn: the message text/images plus
-            end-of-turn tokens. This is the trainable portion.
-            Examples: " Hello world\\n\\n" (RoleColon), "Hello world<|eot_id|>" (Llama3).
-        header: Role identifier and delimiters that introduce the turn. This is what the
-            model sees but does not generate.
-            Examples: "User:" (RoleColon), "<|start_header_id|>user<|end_header_id|>\\n\\n" (Llama3).
-            Typically receives zero training weight.
-        stop_overlap: Edge case field for formats where the stop sequence spans message
-            boundaries. Most renderers (Llama3, Qwen3, DeepSeek, etc.) don't use this—their
-            stop tokens are included in output.
-
-            Only RoleColonRenderer uses this. Its stop sequence is "\\n\\nUser:", where "\\n\\n"
-            ends the output but "User:" would duplicate the next message's header. To avoid
-            duplication, "User:" is stored here and only appended for the last message in
-            supervised training. The name "stop_overlap" reflects that these tokens are the
-            overlap between the stop sequence and the next message's header.
-    """
-
-    output: list[tinker.ModelInputChunk]
-    """What the model generates for this turn."""
-
-    header: tinker.EncodedTextChunk | None = None
-    """Role identifier and delimiters that introduce the turn."""
-
-    stop_overlap: tinker.EncodedTextChunk | None = None
-    """Tokens that overlap between stop sequence and next message's header."""
-
-
-class TrainOnWhat(StrEnum):
-    LAST_ASSISTANT_MESSAGE = "last_assistant_message"
-    ALL_ASSISTANT_MESSAGES = "all_assistant_messages"
-    ALL_MESSAGES = "all_messages"
-    ALL_TOKENS = "all_tokens"
-    ALL_USER_AND_SYSTEM_MESSAGES = "all_user_and_system_messages"
-    CUSTOMIZED = "customized"
-
-
-class Renderer(ABC):
-    """
-    Abstract base class for rendering message lists into training and sampling prompts.
-
-    Subclasses must implement:
-    - get_stop_sequences(): Return stop tokens/strings for sampling
-    - render_message(): Break a message into header/output/stop_overlap components
-    - parse_response(): Convert sampled tokens back into a Message
-
-    The default build_generation_prompt and build_supervised_example implementations
-    assume simple concatenation of rendered messages. Override these if your renderer
-    modifies the conversation structure (e.g., stripping thinking blocks from history).
-    """
-
-    tokenizer: Tokenizer
-
-    def __init__(self, tokenizer: Tokenizer):
-        self.tokenizer = tokenizer
-
-    @property
-    def _bos_tokens(self) -> list[int]:
-        return []
-
-    @abstractmethod
-    def get_stop_sequences(self) -> list[str] | list[int]:
-        """Return the stop sequences used when sampling from this renderer."""
-        ...
-
-    @abstractmethod
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        """
-        Render a single message into its header/output/stop_overlap components.
-
-        This method breaks down a message into parts for loss masking. See RenderedMessage
-        for detailed semantics of each component.
-
-        Args:
-            message: The message to render.
-            ctx: Context about the message's position in the conversation, including:
-                - idx: The index of this message (0-based)
-                - is_last: Whether this is the last message
-                - prev_message: The previous message, if any
-
-        Returns:
-            RenderedMessage with header, output, and optionally stop_overlap.
-        """
-        ...
-
-    @abstractmethod
-    def parse_response(self, response: list[int]) -> tuple[Message, bool]:
-        """
-        Parse sampled tokens back into a Message.
-
-        Args:
-            response: Token IDs returned from sampling.
-
-        Returns:
-            A tuple of (message, success). If success is False, the response could not
-            be parsed (e.g., missing stop token), but a best-effort message is still returned.
-        """
-        ...
-
-    def create_conversation_prefix_with_tools(
-        self, tools: list[ToolSpec], system_prompt: str = ""
-    ) -> list[Message]:
-        """Create message(s) with tool specifications to prepend to conversations.
-
-        Returns one or more messages to prepend to the conversation. This is the
-        standard way to add tools - the returned messages should be placed at the
-        start of your message list before user/assistant messages.
-
-        Args:
-            tools: List of tool specifications.
-            system_prompt: The system prompt content.
-
-        Returns:
-            List of messages to prepend to the conversation.
-
-        Raises:
-            NotImplementedError: If the renderer doesn't support tool calling.
-        """
-        raise NotImplementedError
-
-    def _get_generation_suffix(self, role: Role, ctx: RenderContext) -> list[int]:
-        """Return tokens to append when prompting for generation.
-
-        This is called by build_generation_prompt to add the appropriate prefix
-        before the model starts generating. Override this method instead of
-        build_generation_prompt when you only need to customize the suffix.
-
-        Args:
-            role: The role to generate (usually "assistant")
-            ctx: Context for the generation suffix. Note that ctx.is_last is False
-                because we're prompting for generation, not rendering a complete message.
-
-        Returns:
-            List of token IDs for the role header. Examples in string form:
-            - Llama3: "<|start_header_id|>assistant<|end_header_id|>\n\n"
-            - Qwen3: "<|im_start|>assistant\n"
-            - DeepSeek: "<｜Assistant｜>" (single special token)
-        """
-        # Default: render an empty message and use its header tokens
-        rendered = self.render_message(Message(role=role, content=""), ctx)
-        if rendered.header:
-            return list(rendered.header.tokens)
-        return []
-
-    def build_generation_prompt(
-        self, messages: list[Message], role: Role = "assistant", prefill: str | None = None
-    ) -> tinker.ModelInput:
-        """
-        Generates tokens for sampling from the model.
-
-        Args:
-            messages: a list of messages to render.
-            role: the role of the partial message to be completed.
-            prefill: an optional string to prefill in the model's generation.
-        """
-
-        chunks: list[tinker.types.ModelInputChunk] = []
-        if self._bos_tokens:
-            chunks.append(tinker.types.EncodedTextChunk(tokens=self._bos_tokens))
-        for idx, message in enumerate(messages):
-            ctx = RenderContext(
-                idx=idx,
-                is_last=(idx == len(messages) - 1),
-                prev_message=messages[idx - 1] if idx > 0 else None,
-            )
-            rendered_message = self.render_message(message, ctx)
-            header_chunk = rendered_message.header
-            output_chunks = rendered_message.output
-            if header_chunk:
-                chunks.append(header_chunk)
-            chunks.extend([x for x in output_chunks if x])
-
-        suffix_ctx = RenderContext(
-            idx=len(messages),
-            is_last=True,
-            prev_message=messages[-1] if messages else None,
-        )
-        suffix_tokens = self._get_generation_suffix(role, suffix_ctx)
-        if suffix_tokens:
-            chunks.append(tinker.types.EncodedTextChunk(tokens=suffix_tokens))
-
-        if prefill:
-            chunks.append(
-                tinker.types.EncodedTextChunk(
-                    tokens=self.tokenizer.encode(prefill, add_special_tokens=False)
-                )
-            )
-        return tinker.ModelInput(chunks=chunks)
-
-    def build_supervised_example(
-        self,
-        messages: list[Message],
-        train_on_what: TrainOnWhat = TrainOnWhat.LAST_ASSISTANT_MESSAGE,
-    ) -> tuple[tinker.ModelInput, torch.Tensor]:
-        """
-        Build tokens and per-token weights for supervised fine-tuning.
-
-        This default implementation concatenates rendered messages in order. Override
-        this method if your build_generation_prompt does anything that breaks the simple
-        concatenation assumption—for example, if it strips thinking blocks from history
-        (like Qwen3Renderer), injects default system prompts (like KimiK2Renderer), or
-        otherwise modifies the token sequence.
-
-        The supervised example tokens should match what build_generation_prompt would
-        produce for the same conversation prefix, so the model trains on the same
-        distribution it sees at inference time.
-
-        Args:
-            messages: A list of messages to render.
-            train_on_what: Controls which tokens receive non-zero training weight:
-                - LAST_ASSISTANT_MESSAGE: Only the last assistant message
-                - ALL_ASSISTANT_MESSAGES: All assistant messages
-                - ALL_MESSAGES: All messages (but not headers)
-                - ALL_TOKENS: Everything including headers
-                - ALL_USER_AND_SYSTEM_MESSAGES: User and system messages only
-                - CUSTOMIZED: Use the 'trainable' field on each message
-
-        Returns:
-            A tuple of (model_input, weights) where weights is a 1D tensor with the
-            same length as the total number of tokens.
-        """
-        # TODO: Warn if train_on_what != LAST_ASSISTANT_MESSAGE and the renderer
-        # doesn't satisfy the sequence extension property (e.g., strips thinking from
-        # history). In that case, training on multiple assistant messages requires
-        # separate examples with different token sequences, not a single example.
-        # See docs/rl/sequence-extension.mdx for details.
-
-        model_input_chunks_weights: list[tuple[tinker.types.ModelInputChunk, float]] = []
-        if self._bos_tokens:
-            model_input_chunks_weights.append(
-                (tinker.types.EncodedTextChunk(tokens=self._bos_tokens), 0.0)
-            )
-
-        for idx, message in enumerate(messages):
-            if train_on_what == TrainOnWhat.CUSTOMIZED:
-                assert "trainable" in message, (
-                    "When using CUSTOMIZED train_on_what, each message must have a trainable field: True if loss is applied on this message, False otherwise"
-                )
-            else:
-                assert "trainable" not in message, (
-                    "When using non-CUSTOMIZED train_on_what, each message must not have a trainable field. Either change train_on_what to CUSTOMIZED or remove the trainable field from the message"
-                )
-
-            is_last_message = idx == len(messages) - 1
-            is_assistant = message["role"] == "assistant"
-            is_user_or_system = message["role"] in ["user", "system"]
-
-            # only apply weight to header if train_on_what is ALL_TOKENS
-            ctx = RenderContext(
-                idx=idx,
-                is_last=is_last_message,
-                prev_message=messages[idx - 1] if idx > 0 else None,
-            )
-            rendered_message = self.render_message(message, ctx)
-            header_part = rendered_message.header
-            output_parts = rendered_message.output
-            stop_overlap_part = rendered_message.stop_overlap
-
-            header_weight = int(train_on_what == TrainOnWhat.ALL_TOKENS)
-            if header_part:
-                model_input_chunks_weights += [(header_part, header_weight)]
-
-            match train_on_what:
-                case TrainOnWhat.LAST_ASSISTANT_MESSAGE:
-                    output_has_weight = is_last_message and is_assistant
-                case TrainOnWhat.ALL_ASSISTANT_MESSAGES:
-                    output_has_weight = is_assistant
-                case TrainOnWhat.ALL_MESSAGES:
-                    output_has_weight = True
-                case TrainOnWhat.ALL_TOKENS:
-                    output_has_weight = True
-                case TrainOnWhat.ALL_USER_AND_SYSTEM_MESSAGES:
-                    output_has_weight = is_user_or_system
-                case TrainOnWhat.CUSTOMIZED:
-                    output_has_weight = message.get("trainable", False)
-                case _:
-                    raise ValueError(f"Unknown train_on_what: {train_on_what}")
-
-            model_input_chunks_weights += [
-                (output_part, int(output_has_weight)) for output_part in output_parts if output_part
-            ]
-
-            # stop_overlap completes the stop sequence for formats like RoleColon (e.g., "User:")
-            # Only included for the last message.
-            if is_last_message and stop_overlap_part:
-                model_input_chunks_weights += [(stop_overlap_part, int(output_has_weight))]
-
-        weights_data = [w for chunk, w in model_input_chunks_weights for _ in range(chunk.length)]
-        weights_tensor = torch.tensor(weights_data)
-
-        model_input_chunks = [chunk for chunk, _ in model_input_chunks_weights]
-        return tinker.ModelInput(chunks=model_input_chunks), weights_tensor
-
-
-def tokens_weights_from_strings_weights(
-    strings_weights: list[tuple[str, float]],
-    tokenizer: Tokenizer,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    strings, weights = zip(*strings_weights, strict=True)
-    token_chunks = [tokenizer.encode(s, add_special_tokens=i == 0) for i, s in enumerate(strings)]
-    weights = torch.cat(
-        [torch.full((len(chunk),), w) for chunk, w in zip(token_chunks, weights, strict=True)]
-    )
-    tokens = torch.cat([torch.tensor(chunk) for chunk in token_chunks])
-    assert tokens.dtype == torch.int64
-    return tokens, weights
-
-
-def parse_response_for_stop_token(
-    response: list[int], tokenizer: Tokenizer, stop_token: int
-) -> tuple[Message, bool]:
-    """Parse response for a single stop token.
-
-    We expect a properly rendered response to have exactly one stop token; but it may have zero if e.g. the model
-    ran out of tokens when sampling, which will incur a format error. If there are > 1, there is likely a bug in the
-    sampler and we should error.
-    """
-    emt_count = response.count(stop_token)
-    if emt_count == 0:
-        str_response = tokenizer.decode(response)
-        logger.debug(f"Response is not a valid assistant response: {str_response}")
-        return Message(role="assistant", content=str_response), False
-    elif emt_count == 1:
-        str_response = tokenizer.decode(response[: response.index(stop_token)])
-        return Message(role="assistant", content=str_response), True
-    else:
-        raise ValueError(
-            f"When parsing response, expected to split into 1 or 2 pieces using stop tokens, but got {emt_count}. "
-            "You probably are using the wrong stop tokens when sampling"
-        )
-
-
-class RoleColonRenderer(Renderer):
-    """
-    format like this:
-        User: <content>
-
-        Assistant: <content>
-
-    This is basically the format used by DeepSeek, and similar to the format used by Anthropic,
-    except that they use "Human" instead of "User".
-    """
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        header_str = message["role"].capitalize() + ":"
-        output_str = " " + ensure_text(message["content"]) + "\n\n"
-        # stop_overlap completes the stop sequence "\n\nUser:" for assistant messages.
-        # For non-assistant messages, we use a placeholder that's never actually concatenated.
-        stop_overlap_str = "User:" if message["role"] == "assistant" else "<UNUSED>"
-        header = tinker.types.EncodedTextChunk(
-            tokens=self.tokenizer.encode(header_str, add_special_tokens=False)
-        )
-        output: list[tinker.ModelInputChunk] = [
-            tinker.types.EncodedTextChunk(
-                tokens=self.tokenizer.encode(output_str, add_special_tokens=False)
-            )
-        ]
-        stop_overlap = tinker.types.EncodedTextChunk(
-            tokens=self.tokenizer.encode(stop_overlap_str, add_special_tokens=False)
-        )
-        return RenderedMessage(header=header, output=output, stop_overlap=stop_overlap)
-
-    def get_stop_sequences(self) -> list[str]:
-        return ["\n\nUser:"]
-
-    def parse_response(self, response: list[int]) -> tuple[Message, bool]:
-        str_response = self.tokenizer.decode(response)
-        splitted = str_response.split("\n\nUser:")
-        if len(splitted) == 1:
-            logger.debug(f"Response is not a valid assistant response: {str_response}")
-            return Message(role="assistant", content=str_response.strip()), False
-        elif len(splitted) == 2:
-            before, _after = splitted
-            return Message(role="assistant", content=before.strip()), True
-        else:
-            raise ValueError(
-                f"When parsing response, expected to split into 1 or 2 pieces using stop tokens, but got {len(splitted)}. "
-                "You probably are using the wrong stop tokens when sampling"
-            )
-
-    @property
-    def _bos_tokens(self) -> list[int]:
-        bos_token_str = self.tokenizer.bos_token
-        if bos_token_str is None:
-            return []
-        assert isinstance(bos_token_str, str)
-        return self.tokenizer.encode(bos_token_str, add_special_tokens=False)
-
-    def create_conversation_prefix_with_tools(
-        self, tools: list[ToolSpec], system_prompt: str = ""
-    ) -> list[Message]:
-        raise NotImplementedError("RoleColonRenderer does not support tool calling")
-
-
-class Llama3Renderer(Renderer):
-    """
-    Format like this:
-        <|begin_of_text|><|start_header_id|>system<|end_header_id|>
-
-        You are a helpful AI assistant for travel tips and recommendations<|eot_id|><|start_header_id|>user<|end_header_id|>
-
-        What can you help me with?<|eot_id|><|start_header_id|>assistant<|end_header_id|>
-
-    Note: The HF template prepends "Cutting Knowledge Date: December 2023\\nToday Date: {date}"
-    to system messages. We chose not to do this because it seemed janky. If you want to match
-    the HF template exactly, modify render_message to prepend this info for system messages.
-    """
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        # Determine role for header
-        # Tool responses use "ipython" role in Llama 3 format
-        role = message["role"]
-        if role == "tool":
-            role = "ipython"
-
-        header_str = f"<|start_header_id|>{role}<|end_header_id|>\n\n"
-
-        # Build output content
-        output_str = ensure_text(message["content"])
-
-        # Handle tool calls in assistant messages
-        # Llama 3 format: <function=function_name>{"arg": "value"}</function>
-        if "tool_calls" in message and message["tool_calls"]:
-            tool_call_strs = []
-            for tool_call in message["tool_calls"]:
-                func_name = tool_call.function.name
-                args = tool_call.function.arguments
-                tool_call_strs.append(f"<function={func_name}>{args}</function>")
-            output_str += "".join(tool_call_strs)
-
-        output_str += "<|eot_id|>"
-
-        header = tinker.types.EncodedTextChunk(
-            tokens=self.tokenizer.encode(header_str, add_special_tokens=False)
-        )
-        output: list[tinker.ModelInputChunk] = [
-            tinker.types.EncodedTextChunk(
-                tokens=self.tokenizer.encode(output_str, add_special_tokens=False)
-            )
-        ]
-        return RenderedMessage(header=header, output=output)
-
-    @property
-    def _bos_tokens(self) -> list[int]:
-        return self.tokenizer.encode("<|begin_of_text|>", add_special_tokens=False)
-
-    @property
-    def _end_message_token(self) -> int:
-        (token,) = self.tokenizer.encode("<|eot_id|>", add_special_tokens=False)
-        return token
-
-    def get_stop_sequences(self) -> list[int]:
-        return [self._end_message_token]
-
-    def _parse_llama_tool_calls(
-        self, content: str
-    ) -> tuple[list[ToolCall], list[UnparsedToolCall]]:
-        """Parse tool calls from Llama 3 format.
-
-        Llama 3 uses: <function=function_name>{"arg": "value"}</function>
-
-        Returns:
-            Tuple of (successfully parsed tool calls, failed parses).
-        """
-        tool_calls: list[ToolCall] = []
-        unparsed_tool_calls: list[UnparsedToolCall] = []
-
-        for match in re.finditer(
-            r"<function=(\w+)>(.*?)</function>",
-            content,
-            re.DOTALL,
-        ):
-            raw_text = match.group(0)
-            func_name = match.group(1)
-            args_str = match.group(2).strip()
-            try:
-                # Validate JSON
-                json.loads(args_str)
-                tool_calls.append(
-                    ToolCall(
-                        function=ToolCall.FunctionBody(name=func_name, arguments=args_str),
-                    )
-                )
-            except json.JSONDecodeError as e:
-                unparsed_tool_calls.append(
-                    UnparsedToolCall(raw_text=raw_text, error=f"Invalid JSON: {e}")
-                )
-
-        return tool_calls, unparsed_tool_calls
-
-    def parse_response(self, response: list[int]) -> tuple[Message, bool]:
-        assistant_message, parse_success = parse_response_for_stop_token(
-            response, self.tokenizer, self._end_message_token
-        )
-        if not parse_success:
-            return assistant_message, False
-
-        assert isinstance(assistant_message["content"], str)
-        content = assistant_message["content"]
-
-        # Parse tool calls
-        tool_calls, unparsed_tool_calls = self._parse_llama_tool_calls(content)
-        if tool_calls:
-            assistant_message["tool_calls"] = tool_calls
-        if unparsed_tool_calls:
-            assistant_message["unparsed_tool_calls"] = unparsed_tool_calls
-
-        # Strip all function blocks from content (both parsed and unparsed)
-        if tool_calls or unparsed_tool_calls:
-            content = re.sub(
-                r"\s*<function=\w+>.*?</function>",
-                "",
-                content,
-                flags=re.DOTALL,
-            )
-            assistant_message["content"] = content.strip()
-
-        return assistant_message, True
-
-    def create_conversation_prefix_with_tools(
-        self, tools: list[ToolSpec], system_prompt: str = ""
-    ) -> list[Message]:
-        """Create system message with Llama 3 tool specifications.
-
-        Llama 3.1 supports two tool calling formats. We use the function-tag format
-        which works for custom tools without special environment setup.
-
-        Reference: https://github.com/meta-llama/llama-models/blob/main/models/llama3_1/prompt_format.md
-        """
-        tools_text = ""
-        if tools:
-            tool_lines = "\n\n".join(json.dumps(tool, indent=2) for tool in tools)
-            tools_text = f"""You have access to the following functions:
-
-{tool_lines}
-
-If you choose to call a function, ONLY reply in the following format with no prefix or suffix:
-
-<function=example_function_name>{{"example_name": "example_value"}}</function>
-
-Reminder:
-- Function calls MUST follow the specified format, start with <function= and end with </function>
-- Required parameters MUST be specified
-- Only call one function at a time
-- Put the entire function call reply on one line
-
-"""
-
-        return [Message(role="system", content=tools_text + system_prompt)]
-
-
-class Qwen3Renderer(Renderer):
-    """
-    Renderer for Qwen3 models with thinking enabled.
-
-    This renderer is designed to match HuggingFace's Qwen3 chat template behavior
-    (with enable_thinking=True, which is the default). This ensures compatibility
-    with the OpenAI-compatible /chat/completions endpoint, which uses HF templates.
-
-    Reference: https://huggingface.co/Qwen/Qwen3-8B/blob/main/tokenizer_config.json
-
-    Format:
-        <|im_start|>system
-        You are Qwen, created by Alibaba Cloud.<|im_end|>
-        <|im_start|>user
-        What can you help me with?<|im_end|>
-        <|im_start|>assistant
-        <think>
-        [reasoning content]
-        </think>
-        I can help you with...<|im_end|>
-
-    The default strip_thinking_from_history=True matches HF behavior where thinking
-    blocks are stripped from historical assistant messages in multi-turn conversations.
-    Use strip_thinking_from_history=False for multi-turn RL to get the extension property.
-    """
-
-    def __init__(self, tokenizer: Tokenizer, strip_thinking_from_history: bool = True):
-        """
-        Args:
-            tokenizer: The tokenizer to use for encoding.
-            strip_thinking_from_history: When True (default), strips <think>...</think> blocks
-                from assistant messages in multi-turn history. This matches HuggingFace's
-                Qwen3 chat template behavior. Set to False to preserve thinking in history
-                (useful for multi-turn RL where you need the extension property).
-
-        Note: When strip_thinking_from_history=True, this renderer produces identical
-        tokens to HuggingFace's apply_chat_template with enable_thinking=True.
-
-        See /rl/sequence-extension in the docs for details on how strip_thinking_from_history
-        affects multi-turn RL compute efficiency.
-        """
-        super().__init__(tokenizer)
-        self.strip_thinking_from_history = strip_thinking_from_history
-
-    def _get_qwen_role_for_message(self, message: Message) -> str:
-        """Get the role to use for rendering a message in Qwen format.
-
-        Per HuggingFace Qwen3 chat template, tool messages are rendered with role "user".
-        """
-        role = message["role"]
-        if role == "tool":
-            return "user"
-        return role
-
-    def _wrap_qwen_tool_response(self, content: str) -> str:
-        """Wrap tool response content in Qwen's <tool_response> tags."""
-        return f"<tool_response>\n{content}\n</tool_response>"
-
-    def _should_add_think_prefix(
-        self, message: Message, output_content: str, ctx: RenderContext
-    ) -> bool:
-        """Whether to add <think> prefix to the last assistant message.
-
-        Thinking-enabled models (default) check if this is the last assistant message
-        and if <think> is not already present. Override in subclasses like
-        Qwen3InstructRenderer to disable the <think> prefix entirely.
-        """
-        return message["role"] == "assistant" and "<think>" not in output_content and ctx.is_last
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        maybe_newline = "\n" if ctx.idx > 0 else ""
-
-        role = self._get_qwen_role_for_message(message)
-        header_str = f"{maybe_newline}<|im_start|>{role}\n"
-
-        content = message["content"]
-
-        if isinstance(content, list):
-            # Structured content - handle with list operations
-            parts = content
-            if (
-                self.strip_thinking_from_history
-                and message["role"] == "assistant"
-                and not ctx.is_last
-            ):
-                # Remove thinking parts for historical messages
-                parts = remove_thinking(parts)
-            # Render parts in order, preserving interleaved thinking/text structure.
-            # No separator needed - whitespace is preserved in TextPart for roundtrip identity.
-            rendered_parts = []
-            for p in parts:
-                if p["type"] == "thinking":
-                    rendered_parts.append(f"<think>{p['thinking']}</think>")
-                elif p["type"] == "text":
-                    rendered_parts.append(p["text"])
-                # ToolCallPart handled via message's tool_calls field
-            output_content = "".join(rendered_parts)
-        else:
-            # String content - pass through as-is.
-            # Note: strip_thinking_from_history only works with list-based content.
-            # For stripping to work on historical messages, use structured content
-            # with ThinkingPart separated from text (as returned by parse_response).
-            output_content = content
-
-        # Handle tool response wrapping
-        if message["role"] == "tool":
-            output_content = self._wrap_qwen_tool_response(output_content)
-
-        # Add <think> prefix for last assistant message if not already present
-        if self._should_add_think_prefix(message, output_content, ctx):
-            # Matching the paper, we force the assistant to start with <think>. Some SFT datasets include
-            # <think> in the assistant messages, so we don't need to re-add it in those cases.
-            header_str += "<think>\n"
-
-        # Handle tool_calls field
-        if "tool_calls" in message:
-            # Add leading newline to match HF template behavior
-            output_content += "\n" + "\n".join(
-                [
-                    f"<tool_call>\n{json.dumps(_tool_call_payload(tool_call))}\n</tool_call>"
-                    for tool_call in message["tool_calls"]
-                ]
-            )
-        output_content += "<|im_end|>"
-        header = tinker.types.EncodedTextChunk(
-            tokens=self.tokenizer.encode(header_str, add_special_tokens=False)
-        )
-        output: list[tinker.ModelInputChunk] = [
-            tinker.types.EncodedTextChunk(
-                tokens=self.tokenizer.encode(output_content, add_special_tokens=False)
-            )
-        ]
-        return RenderedMessage(header=header, output=output)
-
-    def _get_generation_suffix(self, role: Role, ctx: RenderContext) -> list[int]:
-        """Return the generation suffix for Qwen3.
-
-        For Qwen3 in thinking mode, we only add the role header (e.g., <|im_start|>assistant\n).
-        We do NOT add <think> here - the model generates that itself. This matches HF's
-        add_generation_prompt=True behavior.
-        """
-        maybe_newline = "\n" if ctx.idx > 0 else ""
-        header_str = f"{maybe_newline}<|im_start|>{role}\n"
-        return self.tokenizer.encode(header_str, add_special_tokens=False)
-
-    @property
-    def _end_message_token(self) -> int:
-        tokens = self.tokenizer.encode("<|im_end|>", add_special_tokens=False)
-        assert len(tokens) == 1, f"Expected single token for <|im_end|>, got {len(tokens)}"
-        return tokens[0]
-
-    def get_stop_sequences(self) -> list[int]:
-        return [self._end_message_token]
-
-    def parse_response(self, response: list[int]) -> tuple[Message, bool]:
-        assistant_message, parse_success = parse_response_for_stop_token(
-            response, self.tokenizer, self._end_message_token
-        )
-        if not parse_success:
-            return assistant_message, False
-
-        # Parse <think>...</think> and <tool_call>...</tool_call> blocks together
-        # to preserve ordering. Tool calls use Qwen's format:
-        # - https://qwen.readthedocs.io/en/latest/getting_started/concepts.html#tool-calling
-        # - https://github.com/QwenLM/Qwen-Agent/blob/main/qwen_agent/llm/fncall_prompts/nous_fncall_prompt.py#L279-L282
-        assert isinstance(assistant_message["content"], str)
-        content = assistant_message["content"]
-
-        # Parse all blocks in one pass, preserving order
-        parts = parse_content_blocks(content)
-
-        if parts is not None:
-            assistant_message["content"] = parts
-
-            # Also populate tool_calls and unparsed_tool_calls fields for backward compatibility
-            # TODO: Consider moving away from TypedDicts for part types - current approach
-            # relies on runtime type checking (p["type"] == "tool_call") without static guarantees.
-            tool_calls = [p["tool_call"] for p in parts if p["type"] == "tool_call"]
-            if tool_calls:
-                assistant_message["tool_calls"] = tool_calls
-
-            unparsed = [
-                UnparsedToolCall(raw_text=p["raw_text"], error=p["error"])
-                for p in parts
-                if p["type"] == "unparsed_tool_call"
-            ]
-            if unparsed:
-                assistant_message["unparsed_tool_calls"] = unparsed
-        else:
-            # No special blocks found - keep as string for backward compatibility
-            assistant_message["content"] = content
-
-        return assistant_message, True
-
-    def create_conversation_prefix_with_tools(
-        self, tools: list[ToolSpec], system_prompt: str = ""
-    ) -> list[Message]:
-        """Create system message with Qwen3 tool specifications.
-
-        Qwen3 uses XML `<tools>` tags with JSON tool definitions appended to the
-        system message content.
-
-        Reference: https://huggingface.co/Qwen/Qwen3-8B/blob/main/tokenizer_config.json
-        """
-        tools_text = ""
-        if tools:
-            tool_lines = "\n".join(json.dumps(tool, separators=(",", ":")) for tool in tools)
-            tools_text = f"""
-
-# Tools
-
-You may call one or more functions to assist with the user query.
-
-You are provided with function signatures within <tools></tools> XML tags:
-<tools>
-{tool_lines}
-</tools>
-
-For each function call, return a json object with function name and arguments within <tool_call></tool_call> XML tags:
-<tool_call>
-{{"name": <function-name>, "arguments": <args-json-object>}}
-</tool_call>"""
-
-        return [Message(role="system", content=system_prompt + tools_text)]
-
-
-class Qwen3DisableThinkingRenderer(Qwen3Renderer):
-    """
-    Renderer for Qwen3 hybrid models with thinking disabled.
-
-    This renderer matches HuggingFace's Qwen3 chat template behavior with
-    enable_thinking=False (or thinking=False for apply_chat_template). It adds
-    empty <think>\\n\\n</think>\\n\\n blocks to assistant messages, signaling to
-    the model that it should respond directly without extended reasoning.
-
-    Use this renderer when you want to train or sample from Qwen3 models in
-    "non-thinking" mode while maintaining compatibility with the OpenAI endpoint.
-    """
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        # Add empty thinking block only to the LAST assistant message (matching HF behavior)
-        if message["role"] == "assistant" and ctx.is_last:
-            content = message.get("content", "")
-            message = message.copy()
-            if isinstance(content, str):
-                if "<think>" not in content:
-                    message["content"] = "<think>\n\n</think>\n\n" + content
-            else:
-                # List content - prepend empty ThinkingPart if no thinking already present
-                has_thinking = any(p["type"] == "thinking" for p in content)
-                if not has_thinking:
-                    message["content"] = [
-                        ThinkingPart(type="thinking", thinking="\n\n"),
-                        *content,
-                    ]
-        return super().render_message(message, ctx)
-
-    def build_generation_prompt(
-        self, messages: list[Message], role: Role = "assistant", prefill: str | None = None
-    ) -> tinker.ModelInput:
-        prefill = "<think>\n\n</think>\n\n" + (prefill or "")
-        return super().build_generation_prompt(messages, role, prefill)
-
-
-class Qwen3InstructRenderer(Qwen3Renderer):
-    """
-    Renderer for Qwen3 instruct 2507 models. Unlike the earlier Qwen3 models, these models do not
-    use the <think> tag at all.
-
-    Inherits from Qwen3Renderer but disables the <think> prefix for assistant messages.
-    ThinkingPart in content is still handled (rendered as <think>...</think>) in case
-    the conversation includes thinking, but no prefix is added automatically.
-    """
-
-    def _should_add_think_prefix(
-        self, message: Message, output_content: str, ctx: RenderContext
-    ) -> bool:
-        """Instruct models don't use <think> prefix."""
-        return False
-
-
-class ImageProcessorProtocol(Protocol):
-    merge_size: int
-    patch_size: int
-
-    def get_number_of_image_patches(
-        self, height: int, width: int, images_kwargs: Optional[dict] = None
-    ) -> int:
-        raise NotImplementedError()
-
-
-def image_to_chunk(
-    image_or_str: Image.Image | str, image_processor: ImageProcessorProtocol
-) -> tinker.types.ImageChunk:
-    """
-    Convert a PIL Image to a tinker.types.ImageChunk for QwenVL
-    """
-
-    # load an image from a data URI or a URL
-    if isinstance(image_or_str, str):
-        with urllib.request.urlopen(image_or_str) as response:
-            pil_image = Image.open(io.BytesIO(response.read()))
-
-    # Otherwise the image is a PIL image and can be loaded directly
-    elif isinstance(image_or_str, Image.Image):
-        pil_image = image_or_str
-
-    # Validate the provided data is actually a valid image type
-    else:
-        raise ValueError("The provided image must be a PIL.Image.Image, URL, or data URI.")
-
-    # Convert to RGB if needed (JPEG doesn't support RGBA/LA/P modes)
-    if pil_image.mode in ("RGBA", "LA", "P"):
-        pil_image = pil_image.convert("RGB")
-
-    img_byte_arr = io.BytesIO()
-    pil_image.save(img_byte_arr, format="JPEG")
-    image_data = img_byte_arr.getvalue()
-
-    width, height = pil_image.size
-    num_image_tokens = (
-        image_processor.get_number_of_image_patches(height, width, images_kwargs={})
-        // image_processor.merge_size**2
-    )
-
-    return tinker.types.ImageChunk(
-        data=image_data,
-        format="jpeg",
-        expected_tokens=num_image_tokens,
-    )
-
-
-class Qwen3VLRenderer(Qwen3Renderer):
-    """
-    Vision-language renderer for Qwen3-VL models with thinking support.
-
-    Format like this:
-        <|im_start|>system
-        You are Qwen, created by Alibaba Cloud. You are a helpful assistant.<|im_end|>
-        <|im_start|>user
-        What can you help me with?<|im_end|>
-        <|im_start|>assistant
-        <think>
-
-        </think>
-        I can help you with...<|im_end|>
-
-    The default strip_thinking_from_history=True matches the non-VL Qwen3Renderer behavior.
-    """
-
-    image_processor: ImageProcessor
-
-    def __init__(
-        self,
-        tokenizer: Tokenizer,
-        image_processor: ImageProcessor,
-        strip_thinking_from_history: bool = True,
-    ):
-        self.tokenizer = tokenizer
-        self.image_processor = image_processor
-        self.strip_thinking_from_history = strip_thinking_from_history
-
-    def _preprocess_message_parts(
-        self, message: Message, *, strip_thinking: bool = False
-    ) -> list[ImagePart | TextPart]:
-        """Convert message content to list form for VL rendering.
-
-        Converts ThinkingPart to <think>...</think> text (or strips if strip_thinking=True).
-        Wraps images with vision tokens. ToolCallPart is not supported in VL content list
-        (use message's tool_calls field instead).
-        """
-        content = message["content"]
-        if isinstance(content, str):
-            base_parts: list[ImagePart | TextPart] = [TextPart(type="text", text=content)]
-        else:
-            # Convert structured content to ImagePart/TextPart list
-            base_parts: list[ImagePart | TextPart] = []
-            for p in content:
-                if p["type"] == "text":
-                    base_parts.append(cast(TextPart, p))
-                elif p["type"] == "image":
-                    base_parts.append(cast(ImagePart, p))
-                elif p["type"] == "thinking":
-                    if not strip_thinking:
-                        # Render thinking as <think>...</think> text
-                        base_parts.append(
-                            TextPart(type="text", text=f"<think>{p['thinking']}</think>")
-                        )
-                    # else: strip thinking by not appending
-                # ToolCallPart and UnparsedToolCallPart are handled via message's tool_calls field
-
-        # Wrap images with vision tokens
-        chunks: list[ImagePart | TextPart] = []
-        for content_chunk in base_parts:
-            if content_chunk["type"] == "image":
-                chunks.append(TextPart(type="text", text="<|vision_start|>"))
-
-            chunks.append(content_chunk)
-
-            if content_chunk["type"] == "image":
-                chunks.append(TextPart(type="text", text="<|vision_end|>"))
-
-        return chunks
-
-    def _wrap_qwen_tool_response_chunks(
-        self, chunks: list[ImagePart | TextPart]
-    ) -> list[ImagePart | TextPart]:
-        """Wrap content chunks in Qwen's <tool_response> tags for multimodal messages."""
-        return (
-            [TextPart(type="text", text="<tool_response>\n")]
-            + chunks
-            + [TextPart(type="text", text="\n</tool_response>")]
-        )
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        maybe_newline = "\n" if ctx.idx > 0 else ""
-
-        role = self._get_qwen_role_for_message(message)
-        header_str = f"{maybe_newline}<|im_start|>{role}\n"
-
-        # Strip thinking from history for non-last assistant messages (matching non-VL behavior)
-        strip_thinking = (
-            self.strip_thinking_from_history and message["role"] == "assistant" and not ctx.is_last
-        )
-        output_chunks = self._preprocess_message_parts(message, strip_thinking=strip_thinking)
-
-        # Handle tool response wrapping
-        if message["role"] == "tool":
-            output_chunks = self._wrap_qwen_tool_response_chunks(output_chunks)
-
-        # Check if any text chunk contains <think> already
-        output_content_for_think_check = "".join(
-            x["text"] for x in output_chunks if isinstance(x, dict) and x["type"] == "text"
-        )
-        if self._should_add_think_prefix(message, output_content_for_think_check, ctx):
-            # Matching the paper, we force the assistant to start with <think>. Some SFT datasets include
-            # <think> in the assistant messages, we so don't need to re-add it in those cases.
-            header_str += "<think>\n"
-        if "tool_calls" in message:
-            output_chunks += [
-                TextPart(
-                    type="text",
-                    text="\n".join(
-                        [
-                            f"<tool_call>\n{json.dumps(_tool_call_payload(tool_call))}\n</tool_call>"
-                            for tool_call in message["tool_calls"]
-                        ]
-                    ),
-                )
-            ]
-        output_chunks += [TextPart(type="text", text="<|im_end|>")]
-
-        output_chunks_encoded: list[tinker.ModelInputChunk] = [
-            image_to_chunk(
-                image_or_str=x["image"],
-                image_processor=cast(ImageProcessorProtocol, self.image_processor),
-            )
-            if x["type"] == "image"
-            else tinker.EncodedTextChunk(
-                tokens=self.tokenizer.encode(x["text"], add_special_tokens=False)
-            )
-            for x in output_chunks
-        ]
-
-        header = tinker.types.EncodedTextChunk(
-            tokens=self.tokenizer.encode(header_str, add_special_tokens=False)
-        )
-        return RenderedMessage(header=header, output=output_chunks_encoded)
-
-
-class Qwen3VLInstructRenderer(Qwen3VLRenderer):
-    """
-    Renderer for Qwen3-VL Instruct models.
-
-    Unlike the Qwen3-VL Thinking models, The Qwen3-VL Instruct models do not use the <think> tag.
-    """
-
-    def _should_add_think_prefix(
-        self, message: Message, output_content: str, ctx: RenderContext
-    ) -> bool:
-        """Instruct models don't use <think> prefix."""
-        return False
-
-
-class DeepSeekV3ThinkingRenderer(Renderer):
-    """
-    Renderer for DeepSeek V3 models in THINKING mode.
-
-    Format:
-        <|begin_of_sentence|><|User|>question<|Assistant|><think>reasoning</think>answer<|end_of_sentence|>
-
-    For non-thinking mode, use DeepSeekV3DisableThinkingRenderer instead.
-
-    System messages at position 0 are rendered without role tokens (matching HF template).
-    System messages at later positions require system_role_as_user=True to convert to user role.
-
-    The default strip_thinking_from_history=True matches HF behavior where thinking
-    traces are removed from historical assistant messages in multi-turn conversations.
-    Use strip_thinking_from_history=False for multi-turn RL to get the extension property.
-    """
-
-    def __init__(
-        self,
-        tokenizer: Tokenizer,
-        system_role_as_user: bool = False,
-        strip_thinking_from_history: bool = True,
-    ):
-        super().__init__(tokenizer)
-        self.system_role_as_user = system_role_as_user
-        self.strip_thinking_from_history = strip_thinking_from_history
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        """Render a single message to tokens.
-
-        Args:
-            message: The message to render.
-            ctx: Context about the message's position, including:
-                - idx: The index of this message (0-based)
-                - is_last: Whether this is the last message (affects thinking stripping)
-                - prev_message: The previous message, used to detect post-tool formatting
-        """
-        # Check if this assistant message follows a tool response
-        follows_tool = ctx.prev_message is not None and ctx.prev_message["role"] == "tool"
-
-        content = message["content"]
-
-        if message["role"] == "system":
-            # HF template collects all system messages at the start without role tokens
-            # We only support this for idx=0; later system messages need system_role_as_user=True
-            content_str = ensure_text(content)
-            if ctx.idx == 0:
-                header_tokens: list[int] = []
-                output_str = content_str
-            elif self.system_role_as_user:
-                # Convert later system messages to user role
-                role_token = self._get_special_token("User")
-                header_tokens = [role_token]
-                output_str = content_str
-            else:
-                raise ValueError(
-                    "DeepSeek only supports system message at start. "
-                    "Use system_role_as_user=True to convert later system messages to user role."
-                )
-        elif message["role"] == "user":
-            role_token = self._get_special_token("User")
-            header_tokens = [role_token]
-            output_str = ensure_text(content)
-        elif message["role"] == "assistant":
-            has_tool_calls = "tool_calls" in message and message["tool_calls"]
-
-            if isinstance(content, list):
-                # Structured content - handle with list operations
-                parts = content
-                if self.strip_thinking_from_history and not has_tool_calls and not ctx.is_last:
-                    # Remove thinking parts for historical messages
-                    parts = remove_thinking(parts)
-                # Render parts in order, preserving interleaved thinking/text structure.
-                # No separator needed - whitespace is preserved in TextPart for roundtrip identity.
-                rendered_parts = []
-                for p in parts:
-                    if p["type"] == "thinking":
-                        rendered_parts.append(f"<think>{p['thinking']}</think>")
-                    elif p["type"] == "text":
-                        rendered_parts.append(p["text"])
-                    # ToolCallPart handled via message's tool_calls field
-                output_content = "".join(rendered_parts)
-            else:
-                # String content - pass through as-is.
-                # Note: strip_thinking_from_history only works with list-based content.
-                # For stripping to work on historical messages, use structured content
-                # with ThinkingPart separated from text (as returned by parse_response).
-                output_content = content
-
-            if follows_tool:
-                # Post-tool assistant: no role token, content flows directly after tool output
-                header_tokens = []
-                output_str = output_content
-            else:
-                # Normal assistant message
-                role_token = self._get_special_token("Assistant")
-                header_tokens = [role_token]
-                output_str = output_content
-        elif message["role"] == "tool":
-            # Tool responses use special tool output tokens to match HF template
-            header_tokens = self.tokenizer.encode(
-                "<｜tool▁output▁begin｜>", add_special_tokens=False
-            )
-            output_str = ensure_text(content) + "<｜tool▁output▁end｜>"
-        else:
-            raise ValueError(f"Unsupported role: {message['role']}")
-
-        # Handle tool calls in assistant messages
-        # HF format: <｜tool▁calls▁begin｜><｜tool▁call▁begin｜>name<｜tool▁sep｜>args<｜tool▁call▁end｜><｜tool▁calls▁end｜>
-        if "tool_calls" in message and message["tool_calls"]:
-            output_str += "<｜tool▁calls▁begin｜>"
-            for tool_call in message["tool_calls"]:
-                func_name = tool_call.function.name
-                args = tool_call.function.arguments
-                output_str += (
-                    f"<｜tool▁call▁begin｜>{func_name}<｜tool▁sep｜>{args}<｜tool▁call▁end｜>"
-                )
-            output_str += "<｜tool▁calls▁end｜>"
-
-        output_tokens = self.tokenizer.encode(output_str, add_special_tokens=False)
-
-        # Add end_of_sentence only for assistant messages with content
-        # (not for empty generation prompt messages)
-        if message["role"] == "assistant" and message["content"]:
-            output_tokens.append(self._end_message_token)
-
-        output: list[tinker.ModelInputChunk] = [tinker.types.EncodedTextChunk(tokens=output_tokens)]
-        # Only include header if non-empty; tinker rejects empty token chunks with
-        # "Chunk N has empty tokens list". This happens for system messages at idx=0.
-        if header_tokens:
-            return RenderedMessage(
-                header=tinker.types.EncodedTextChunk(tokens=header_tokens), output=output
-            )
-        else:
-            return RenderedMessage(output=output)
-
-    def _get_special_token(self, name: str) -> int:
-        sep = chr(65372)
-        s = f"<{sep}{name}{sep}>"
-        res = self.tokenizer.encode(s, add_special_tokens=False)
-        assert len(res) == 1, f"Expected single token for {s}, got {res}"
-        return res[0]
-
-    @property
-    def _bos_tokens(self) -> list[int]:
-        return [self._get_special_token("begin▁of▁sentence")]
-
-    @property
-    def _end_message_token(self) -> int:
-        return self._get_special_token("end▁of▁sentence")
-
-    def get_stop_sequences(self) -> list[int]:
-        return [self._end_message_token]
-
-    def _parse_deepseek_tool_calls(
-        self, content: str
-    ) -> tuple[list[ToolCall], list[UnparsedToolCall]]:
-        """Parse tool calls from DeepSeek V3.1 format.
-
-        Expected format (per HuggingFace model card and chat template):
-            <｜tool▁calls▁begin｜><｜tool▁call▁begin｜>func_name<｜tool▁sep｜>{"arg":"value"}<｜tool▁call▁end｜><｜tool▁calls▁end｜>
-
-        Multiple tool calls are chained directly without separators.
-
-        References:
-            - DeepSeek V3.1 Model Card: https://huggingface.co/deepseek-ai/DeepSeek-V3.1
-            - Chat Template: https://huggingface.co/deepseek-ai/DeepSeek-V3.1/blob/main/assets/chat_template.jinja
-        """
-        tool_calls: list[ToolCall] = []
-        unparsed_tool_calls: list[UnparsedToolCall] = []
-
-        calls_match = re.search(
-            r"<｜tool▁calls▁begin｜>(.*?)<｜tool▁calls▁end｜>", content, re.DOTALL
-        )
-        if not calls_match:
-            return tool_calls, unparsed_tool_calls
-
-        for match in re.finditer(
-            r"<｜tool▁call▁begin｜>(\w+)<｜tool▁sep｜>(.*?)<｜tool▁call▁end｜>",
-            calls_match.group(1),
-            re.DOTALL,
-        ):
-            raw_text = match.group(0)
-            func_name, args_str = match.group(1), match.group(2).strip()
-
-            try:
-                json.loads(args_str)
-                tool_calls.append(
-                    ToolCall(function=ToolCall.FunctionBody(name=func_name, arguments=args_str))
-                )
-            except json.JSONDecodeError as e:
-                unparsed_tool_calls.append(
-                    UnparsedToolCall(raw_text=raw_text, error=f"Invalid JSON: {e}")
-                )
-
-        return tool_calls, unparsed_tool_calls
-
-    def parse_response(self, response: list[int]) -> tuple[Message, bool]:
-        assistant_message, parse_success = parse_response_for_stop_token(
-            response, self.tokenizer, self._end_message_token
-        )
-        if not parse_success:
-            return assistant_message, False
-
-        assert isinstance(assistant_message["content"], str)
-        content = assistant_message["content"]
-
-        # Parse DeepSeek-specific tool calls
-        tool_calls, unparsed_tool_calls = self._parse_deepseek_tool_calls(content)
-        if tool_calls:
-            assistant_message["tool_calls"] = tool_calls
-        if unparsed_tool_calls:
-            assistant_message["unparsed_tool_calls"] = unparsed_tool_calls
-
-        # Strip tool calls section from content (both parsed and unparsed)
-        if tool_calls or unparsed_tool_calls:
-            content = re.sub(
-                r"\s*<｜tool▁calls▁begin｜>.*?<｜tool▁calls▁end｜>",
-                "",
-                content,
-                flags=re.DOTALL,
-            )
-            content = content.strip()
-
-        # Parse <think>...</think> blocks into ThinkingPart/TextPart list
-        parts = parse_think_blocks(content)
-        if parts is not None:
-            assistant_message["content"] = parts
-        else:
-            assistant_message["content"] = content
-
-        return assistant_message, True
-
-    def create_conversation_prefix_with_tools(
-        self, tools: list[ToolSpec], system_prompt: str = ""
-    ) -> list[Message]:
-        """Create system message with DeepSeek V3.1 tool specifications.
-
-        DeepSeek V3.1 tool calling requires tools to be described in the system message
-        using a specific format with ### headers and inline JSON parameters.
-
-        Note: Tool calling is supported in non-thinking mode only.
-
-        References:
-            - DeepSeek V3.1 Model Card (ToolCall section): https://huggingface.co/deepseek-ai/DeepSeek-V3.1
-            - DeepSeek V3.1 Chat Template: https://huggingface.co/deepseek-ai/DeepSeek-V3.1/blob/main/assets/chat_template.jinja
-            - DeepSeek API Tool Calls Guide: https://api-docs.deepseek.com/guides/tool_calls
-        """
-        tools_text = ""
-        if tools:
-            # Format each tool with ### header, description, and parameters
-            tool_blocks = []
-            for tool in tools:
-                tool_block = f"""### {tool["name"]}
-Description: {tool["description"]}
-
-Parameters: {json.dumps(tool["parameters"])}"""
-                tool_blocks.append(tool_block)
-
-            tools_text = f"""
-
-## Tools
-You have access to the following tools:
-
-{chr(10).join(tool_blocks)}
-
-IMPORTANT: ALWAYS adhere to this exact format for tool use:
-<｜tool▁calls▁begin｜><｜tool▁call▁begin｜>tool_call_name<｜tool▁sep｜>tool_call_arguments<｜tool▁call▁end｜><｜tool▁calls▁end｜>
-
-Where:
-- `tool_call_name` must be an exact match to one of the available tools
-- `tool_call_arguments` must be valid JSON that strictly follows the tool's Parameters Schema
-- For multiple tool calls, chain them directly without separators or spaces"""
-
-        return [Message(role="system", content=system_prompt + tools_text)]
-
-
-class DeepSeekV3DisableThinkingRenderer(DeepSeekV3ThinkingRenderer):
-    """
-    Renderer for DeepSeek V3 models in NON-THINKING mode.
-
-    Format:
-        <|begin_of_sentence|><|User|>question<|Assistant|></think>answer<|end_of_sentence|>
-
-    The </think> prefix signals to the model to skip reasoning and respond directly.
-    Any <think>...</think> blocks in the content are stripped.
-
-    For thinking mode, use DeepSeekV3ThinkingRenderer instead.
-    """
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        """Render message in non-thinking mode.
-
-        For assistant messages (not following tool):
-        - Strip any ThinkingPart from structured content
-        - Prepend </think> to signal non-thinking mode
-        """
-        # Check if this assistant message follows a tool response
-        follows_tool = ctx.prev_message is not None and ctx.prev_message["role"] == "tool"
-
-        if message["role"] == "assistant" and not follows_tool:
-            content = message["content"]
-
-            # Strip thinking from content
-            if isinstance(content, list):
-                # Remove ThinkingPart, keep only text
-                text_content = "".join(p["text"] for p in content if p["type"] == "text")
-            else:
-                # Strip <think>...</think> blocks from string content
-                text_content = re.sub(r"<think>.*?</think>", "", content, flags=re.DOTALL)
-
-            # Prepend </think> to signal non-thinking mode
-            message = message.copy()
-            message["content"] = "</think>" + text_content
-
-        return super().render_message(message, ctx)
-
-    def _get_generation_suffix(self, role: Role, ctx: RenderContext) -> list[int]:
-        """Return <｜Assistant｜></think> for generation, or empty after tool messages."""
-        # No suffix after tool messages - content flows directly
-        if ctx.prev_message is not None and ctx.prev_message["role"] == "tool":
-            return []
-        # Otherwise: <｜Assistant｜></think>
-        role_token = self._get_special_token("Assistant")
-        think_close = self.tokenizer.encode("</think>", add_special_tokens=False)
-        return [role_token] + think_close
-
-
-class KimiK2Renderer(Renderer):
-    """
-    Format for moonshotai/Kimi-K2-Thinking:
-        <|im_system|>system<|im_middle|>You are Kimi, an AI assistant created by Moonshot AI.<|im_end|>
-        <|im_user|>user<|im_middle|>What can you help me with?<|im_end|>
-        <|im_assistant|>assistant<|im_middle|><think>reasoning</think>I can help you with...<|im_end|>
-
-    Historical assistant messages use empty <think></think> blocks, while the final assistant
-    response preserves reasoning_content in the thinking block.
-
-    Note: Per the HuggingFace chat template, the default system message is automatically
-    prepended if no system message is provided. This ensures train-eval consistency when
-    using HF's apply_chat_template for inference.
-    """
-
-    DEFAULT_SYSTEM_PROMPT = "You are Kimi, an AI assistant created by Moonshot AI."
-
-    def _ensure_system_message(self, messages: list[Message]) -> list[Message]:
-        """Prepend default system message if no system message is present.
-
-        This matches the HuggingFace chat template behavior where a default system
-        message is automatically added when none is provided.
-        """
-        if not messages or messages[0]["role"] != "system":
-            default_system = Message(role="system", content=self.DEFAULT_SYSTEM_PROMPT)
-            return [default_system] + list(messages)
-        return messages
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        """
-        Render a message. For assistant messages, ctx.is_last controls whether thinking is preserved
-        (True) or stripped to empty <think></think> (False).
-        """
-        role = message["role"]
-        role_name = message.get("name", role)
-
-        # Build role token based on role type
-        if role == "user":
-            header_str = f"<|im_user|>{role_name}<|im_middle|>"
-        elif role == "assistant":
-            header_str = f"<|im_assistant|>{role_name}<|im_middle|>"
-        elif role == "system":
-            header_str = f"<|im_system|>{role_name}<|im_middle|>"
-        elif role == "tool":
-            header_str = f"<|im_system|>{role_name}<|im_middle|>"
-            # Tool responses have special formatting
-            tool_call_id = message.get("tool_call_id", "")
-            header_str += f"## Return of {tool_call_id}\n"
-        else:
-            header_str = f"<|im_system|>{role_name}<|im_middle|>"
-
-        # Build output content
-        output_str = ""
-        if role == "assistant":
-            # Extract thinking and text from content list
-            parts = ensure_list(message["content"])
-            thinking_content = "".join(p["thinking"] for p in parts if p["type"] == "thinking")
-            text_content = "".join(p["text"] for p in parts if p["type"] == "text")
-
-            # For the last assistant message (is_last=True), preserve thinking; otherwise use empty think block
-            if ctx.is_last and thinking_content:
-                output_str = f"<think>{thinking_content}</think>"
-            else:
-                output_str = "<think></think>"
-            output_str += text_content
-
-            # Handle tool calls
-            if "tool_calls" in message and message["tool_calls"]:
-                output_str += "<|tool_calls_section_begin|>"
-                for tool_call in message["tool_calls"]:
-                    tool_id = tool_call.id or ""
-                    args = tool_call.function.arguments
-                    output_str += f"<|tool_call_begin|>{tool_id}<|tool_call_argument_begin|>{args}<|tool_call_end|>"
-                output_str += "<|tool_calls_section_end|>"
-        else:
-            output_str = ensure_text(message["content"])
-
-        output_str += "<|im_end|>"
-
-        header = tinker.types.EncodedTextChunk(tokens=self.tokenizer.encode(header_str))
-        output: list[tinker.ModelInputChunk] = [
-            tinker.types.EncodedTextChunk(tokens=self.tokenizer.encode(output_str))
-        ]
-        return RenderedMessage(header=header, output=output)
-
-    def build_generation_prompt(
-        self, messages: list[Message], role: Role = "assistant", prefill: str | None = None
-    ) -> tinker.ModelInput:
-        messages = self._ensure_system_message(messages)
-        chunks: list[tinker.types.ModelInputChunk] = []
-
-        for idx, message in enumerate(messages):
-            # For generation prompt, no message is "last assistant" since we're generating new response
-            ctx = RenderContext(
-                idx=idx,
-                is_last=False,
-                prev_message=messages[idx - 1] if idx > 0 else None,
-            )
-            rendered_message = self.render_message(message, ctx)
-            header_chunk = rendered_message.header
-            output_chunks = rendered_message.output
-            if header_chunk:
-                chunks.append(header_chunk)
-            chunks.extend([x for x in output_chunks if x])
-
-        # Add generation prompt for new assistant message
-        gen_prompt = f"<|im_assistant|>{role}<|im_middle|>"
-        chunks.append(tinker.types.EncodedTextChunk(tokens=self.tokenizer.encode(gen_prompt)))
-        if prefill:
-            chunks.append(tinker.types.EncodedTextChunk(tokens=self.tokenizer.encode(prefill)))
-        return tinker.ModelInput(chunks=chunks)
-
-    def build_supervised_example(
-        self,
-        messages: list[Message],
-        train_on_what: TrainOnWhat = TrainOnWhat.LAST_ASSISTANT_MESSAGE,
-    ) -> tuple[tinker.ModelInput, torch.Tensor]:
-        """
-        Override to properly handle thinking preservation for the last assistant message.
-        Also ensures default system message is prepended if none is present.
-        """
-        messages = self._ensure_system_message(messages)
-
-        # Find last non-tool-call assistant message index
-        last_assistant_idx = -1
-        for idx in range(len(messages) - 1, -1, -1):
-            if messages[idx]["role"] == "assistant" and "tool_calls" not in messages[idx]:
-                last_assistant_idx = idx
-                break
-
-        model_input_chunks_weights: list[tuple[tinker.types.ModelInputChunk, float]] = []
-
-        for idx, message in enumerate(messages):
-            if train_on_what == TrainOnWhat.CUSTOMIZED:
-                assert "trainable" in message, (
-                    "When using CUSTOMIZED train_on_what, each message must have a trainable field"
-                )
-            else:
-                assert "trainable" not in message, (
-                    "When using non-CUSTOMIZED train_on_what, each message must not have a trainable field"
-                )
-
-            is_last_message = idx == len(messages) - 1
-            is_assistant = message["role"] == "assistant"
-            is_user_or_system = message["role"] in ["user", "system"]
-
-            # For Kimi K2, preserve thinking only for last non-tool-call assistant
-            is_last_assistant = idx >= last_assistant_idx and is_assistant
-            ctx = RenderContext(
-                idx=idx,
-                is_last=is_last_assistant,
-                prev_message=messages[idx - 1] if idx > 0 else None,
-            )
-            rendered_message = self.render_message(message, ctx)
-
-            header_part = rendered_message.header
-            output_parts = rendered_message.output
-
-            header_weight = int(train_on_what == TrainOnWhat.ALL_TOKENS)
-            if header_part:
-                model_input_chunks_weights += [(header_part, header_weight)]
-
-            match train_on_what:
-                case TrainOnWhat.LAST_ASSISTANT_MESSAGE:
-                    output_has_weight = is_last_message and is_assistant
-                case TrainOnWhat.ALL_ASSISTANT_MESSAGES:
-                    output_has_weight = is_assistant
-                case TrainOnWhat.ALL_MESSAGES:
-                    output_has_weight = True
-                case TrainOnWhat.ALL_TOKENS:
-                    output_has_weight = True
-                case TrainOnWhat.ALL_USER_AND_SYSTEM_MESSAGES:
-                    output_has_weight = is_user_or_system
-                case TrainOnWhat.CUSTOMIZED:
-                    output_has_weight = message.get("trainable", False)
-                case _:
-                    raise ValueError(f"Unknown train_on_what: {train_on_what}")
-
-            model_input_chunks_weights += [
-                (output_part, int(output_has_weight)) for output_part in output_parts if output_part
-            ]
-
-        weights_data = [w for chunk, w in model_input_chunks_weights for _ in range(chunk.length)]
-        weights_tensor = torch.tensor(weights_data)
-
-        model_input_chunks = [chunk for chunk, _ in model_input_chunks_weights]
-        return tinker.ModelInput(chunks=model_input_chunks), weights_tensor
-
-    @property
-    def _end_message_token(self) -> int:
-        tokens = self.tokenizer.encode("<|im_end|>")
-        assert len(tokens) == 1, f"Expected single token for <|im_end|>, got {len(tokens)}"
-        return tokens[0]
-
-    def get_stop_sequences(self) -> list[int]:
-        return [self._end_message_token]
-
-    def parse_response(self, response: list[int]) -> tuple[Message, bool]:
-        assistant_message, parse_success = parse_response_for_stop_token(
-            response, self.tokenizer, self._end_message_token
-        )
-        if not parse_success:
-            return assistant_message, False
-
-        content = assistant_message["content"]
-        assert isinstance(content, str)
-
-        # Extract thinking content if present
-        think_match = re.search(r"<think>(.*?)</think>", content, re.DOTALL)
-        if think_match:
-            thinking = think_match.group(1)
-            # Remove the think block from content
-            remaining_content = content[think_match.end() :].lstrip()
-            # Return content as list with ThinkingPart and TextPart
-            content_parts: list[ContentPart] = []
-            if thinking:  # Omit empty thinking
-                content_parts.append(ThinkingPart(type="thinking", thinking=thinking))
-            content_parts.append(TextPart(type="text", text=remaining_content))
-            assistant_message["content"] = content_parts
-            content = remaining_content  # Update content for tool call parsing below
-
-        # Handle tool calls if present
-        if "<|tool_calls_section_begin|>" in content:
-            tool_section_match = re.search(
-                r"<\|tool_calls_section_begin\|>(.*?)<\|tool_calls_section_end\|>",
-                content,
-                re.DOTALL,
-            )
-            if tool_section_match:
-                tool_section = tool_section_match.group(1)
-                tool_calls: list[ToolCall] = []
-                unparsed_tool_calls: list[UnparsedToolCall] = []
-
-                # Parse individual tool calls
-                # Tool ID format: "functions.{func_name}:{idx}" e.g. "functions.get_weather:0"
-                tool_call_pattern = r"<\|tool_call_begin\|>(.*?)<\|tool_call_argument_begin\|>(.*?)<\|tool_call_end\|>"
-                for match in re.finditer(tool_call_pattern, tool_section, re.DOTALL):
-                    raw_text = match.group(0)
-                    tool_id = match.group(1).strip()
-                    args_str = match.group(2).strip()
-
-                    # Extract function name from tool_id (format: "functions.{name}:{idx}")
-                    func_name = ""
-                    if tool_id and "." in tool_id:
-                        # Split on first dot to get "functions" prefix and the rest
-                        parts = tool_id.split(".", 1)
-                        if len(parts) == 2:
-                            # Split on colon to separate name from index
-                            name_part = parts[1].split(":")[0] if ":" in parts[1] else parts[1]
-                            func_name = name_part
-
-                    # Try to parse as JSON to validate, but store as string
-                    try:
-                        json.loads(args_str)
-                        tool_calls.append(
-                            ToolCall(
-                                function=ToolCall.FunctionBody(name=func_name, arguments=args_str),
-                                id=tool_id if tool_id else None,
-                            )
-                        )
-                    except json.JSONDecodeError as e:
-                        unparsed_tool_calls.append(
-                            UnparsedToolCall(raw_text=raw_text, error=f"Invalid JSON: {e}")
-                        )
-
-                if tool_calls:
-                    assistant_message["tool_calls"] = tool_calls
-                if unparsed_tool_calls:
-                    assistant_message["unparsed_tool_calls"] = unparsed_tool_calls
-
-                # Remove tool section from content (both parsed and unparsed)
-                if tool_calls or unparsed_tool_calls:
-                    content = content[: content.find("<|tool_calls_section_begin|>")]
-                    assistant_message["content"] = content
-
-        return assistant_message, True
-
-    def create_conversation_prefix_with_tools(
-        self, tools: list[ToolSpec], system_prompt: str = ""
-    ) -> list[Message]:
-        """Create system messages with Kimi K2 tool specifications.
-
-        Per the HuggingFace chat template, Kimi K2 places the tool_declare message
-        BEFORE the regular system message. If no system_prompt is provided, uses
-        the default system prompt to match HuggingFace chat template behavior.
-
-        Reference: https://huggingface.co/moonshotai/Kimi-K2-Thinking/blob/main/chat_template.jinja
-        """
-        messages: list[Message] = []
-
-        # Tool declaration message comes first (per HF chat template)
-        if tools:
-            tools_json = json.dumps(tools, separators=(",", ":"))
-            messages.append(Message(role="system", content=tools_json, name="tool_declare"))
-
-        # Regular system message second (use default if none provided)
-        actual_system_prompt = system_prompt if system_prompt else self.DEFAULT_SYSTEM_PROMPT
-        messages.append(Message(role="system", content=actual_system_prompt))
-
-        return messages
-
-
-class GptOssRenderer(Renderer):
-    """
-    Format like this (no newlines between messages, last message should end with <|return|> but be replaced by <|end|> when continuing the convo):
-        <|start|>system<|message|>You are ChatGPT...<|end|><|start|>user<|message|>How much is 1+1?<|end|><|start|>assistant<|channel|>final<|message|>2<|end|><|start|>
-    TODO: support channels in input messages and tools
-    """
-
-    system_prompt = "<|start|>system<|message|>You are ChatGPT, a large language model trained by OpenAI.\nKnowledge cutoff: 2024-06\nCurrent date: {current_date}\n\nReasoning: {reasoning_effort}\n\n# Valid channels: analysis, commentary, final. Channel must be included for every message.<|end|>"
-    use_system_prompt: bool = False
-    reasoning_effort: str | None = None
-    current_date: str | None = (
-        None  # If use_system_prompt=True, will use the current date if this is None. Set this to a fixed date for deterministic system prompt.
-    )
-
-    def __init__(
-        self,
-        tokenizer: Tokenizer,
-        use_system_prompt: bool = False,
-        reasoning_effort: str | None = None,
-        current_date: str | None = None,
-    ):
-        super().__init__(tokenizer)
-        self.use_system_prompt = use_system_prompt
-        self.reasoning_effort = reasoning_effort
-        self.current_date = current_date
-        assert use_system_prompt == (reasoning_effort is not None), (
-            "Reasoning effort must be set iff using system prompt"
-        )
-
-    def render_message(self, message: Message, ctx: RenderContext) -> RenderedMessage:
-        assert message.get("tool_calls") is None, "TODO: support tools in gpt-oss renderer"
-        # HF template maps "system" role to "developer" with special formatting
-        role = message["role"]
-        if role == "system":
-            role = "developer"
-        header_str = f"<|start|>{role}"
-        output_str = ""
-        if message["role"] == "assistant":
-            # TODO: support commentary channel / tools
-
-            # Assistant channels. See https://cookbook.openai.com/articles/openai-harmony
-            # Extract text and thinking from content list
-            parts = ensure_list(message["content"])
-            text_content = "".join(p["text"] for p in parts if p["type"] == "text")
-            thinking_content = "".join(p["thinking"] for p in parts if p["type"] == "thinking")
-
-            # Analysis channel (CoT) - always included for last message to match HF template
-            if ctx.is_last:
-                output_str += (
-                    f"<|channel|>analysis<|message|>{thinking_content}<|end|><|start|>assistant"
-                )
-
-            # Final channel (Response Content)
-            output_str += f"<|channel|>final<|message|>{text_content}"
-        elif message["role"] == "system":
-            # HF wraps system content as developer instructions
-            output_str += f"<|message|># Instructions\n\n{ensure_text(message['content'])}\n\n"
-        else:
-            output_str += f"<|message|>{ensure_text(message['content'])}"
-
-        if ctx.is_last and message["role"] == "assistant":
-            # <|return|> is the stop token for assistant generation
-            output_str += "<|return|>"
-        else:
-            output_str += "<|end|>"
-
-        header = tinker.types.EncodedTextChunk(
-            tokens=self.tokenizer.encode(header_str, add_special_tokens=False)
-        )
-        output: list[tinker.ModelInputChunk] = [
-            tinker.types.EncodedTextChunk(
-                tokens=self.tokenizer.encode(output_str, add_special_tokens=False)
-            )
-        ]
-        return RenderedMessage(header=header, output=output)
-
-    def _build_system_prompt(self) -> str:
-        current_date = (
-            self.current_date
-            if self.current_date is not None
-            else datetime.now().strftime("%Y-%m-%d")
-        )
-        return self.system_prompt.format(
-            current_date=current_date, reasoning_effort=self.reasoning_effort
-        )
-
-    @property
-    def _bos_tokens(self) -> list[int]:
-        tokens = []
-        if self.use_system_prompt:
-            tokens.extend(
-                self.tokenizer.encode(self._build_system_prompt(), add_special_tokens=False)
-            )
-        return tokens
-
-    @property
-    def _return_token(self) -> int:
-        res = self.tokenizer.encode("<|return|>", add_special_tokens=False)
-        assert len(res) == 1, f"Expected single token for <|return|>, got {len(res)}"
-        return res[0]
-
-    def get_stop_sequences(self) -> list[int]:
-        return [self._return_token]
-
-    def parse_response(self, response: list[int]) -> tuple[Message, bool]:
-        assistant_message, parse_success = parse_response_for_stop_token(
-            response, self.tokenizer, self._return_token
-        )
-        if not parse_success:
-            return assistant_message, False
-
-        assert isinstance(assistant_message["content"], str)
-        content = assistant_message["content"]
-
-        # Parse GptOss multi-channel format into content parts
-        # Format: <|channel|>channel_name<|message|>content<|end|> or <|return|>
-        # Channels: analysis (thinking), commentary (tool calls), final (response)
-        parts = self._parse_gptoss_channels(content)
-        if parts:
-            assistant_message["content"] = parts
-        # else: keep as string for backward compatibility
-
-        return assistant_message, True
-
-    def _parse_gptoss_channels(self, content: str) -> list[ContentPart]:
-        """Parse GptOss channel format into ContentPart list.
-
-        Channels:
-        - analysis: Chain-of-thought (maps to ThinkingPart)
-        - final: User-facing answer (maps to TextPart)
-        - commentary: Tool calls or preambles (maps to TextPart for now)
-        """
-        if "<|channel|>" not in content:
-            return []
-
-        parts: list[ContentPart] = []
-
-        # Pattern to match channel messages
-        # <|channel|>channel_name<|message|>content<|end|> or <|return|>
-        pattern = re.compile(
-            r"<\|channel\|>(\w+)(?:[^<]*)?<\|message\|>(.*?)(?:<\|end\|>|<\|return\|>|$)",
-            re.DOTALL,
-        )
-
-        for match in pattern.finditer(content):
-            channel = match.group(1)
-            msg_content = match.group(2)
-
-            if not msg_content.strip():
-                continue
-
-            if channel == "analysis":
-                parts.append(ThinkingPart(type="thinking", thinking=msg_content))
-            elif channel == "final":
-                parts.append(TextPart(type="text", text=msg_content))
-            elif channel == "commentary":
-                # Commentary without tool recipient is user-visible preamble
-                # (tool calls would need additional parsing of to=functions.x)
-                parts.append(TextPart(type="text", text=msg_content))
-
-        return parts
-
-    def create_conversation_prefix_with_tools(
-        self, tools: list[ToolSpec], system_prompt: str = ""
-    ) -> list[Message]:
-        raise NotImplementedError("GptOssRenderer does not support tool calling")
-
-
-def get_renderer(
-    name: str, tokenizer: Tokenizer, image_processor: ImageProcessor | None = None
-) -> Renderer:
-    if name == "role_colon":
-        return RoleColonRenderer(tokenizer)
-    elif name == "llama3":
-        return Llama3Renderer(tokenizer)
-    elif name == "qwen3":
-        return Qwen3Renderer(tokenizer)
-    elif name == "qwen3_vl":
-        assert image_processor is not None, "qwen3_vl renderer requires an image_processor"
-        return Qwen3VLRenderer(tokenizer, image_processor)
-    elif name == "qwen3_vl_instruct":
-        assert image_processor is not None, "qwen3_vl_instruct renderer requires an image_processor"
-        return Qwen3VLInstructRenderer(tokenizer, image_processor)
-    elif name == "qwen3_disable_thinking":
-        return Qwen3DisableThinkingRenderer(tokenizer)
-    elif name == "qwen3_instruct":
-        return Qwen3InstructRenderer(tokenizer)
-    elif name == "deepseekv3":
-        # Default to non-thinking mode (matches HF template default behavior)
-        return DeepSeekV3DisableThinkingRenderer(tokenizer)
-    elif name == "deepseekv3_disable_thinking":
-        # Alias for backward compatibility
-        return DeepSeekV3DisableThinkingRenderer(tokenizer)
-    elif name == "deepseekv3_thinking":
-        return DeepSeekV3ThinkingRenderer(tokenizer)
-    elif name == "kimi_k2":
-        return KimiK2Renderer(tokenizer)
-    elif name == "gpt_oss_no_sysprompt":
-        return GptOssRenderer(tokenizer, use_system_prompt=False)
-    elif name == "gpt_oss_low_reasoning":
-        return GptOssRenderer(tokenizer, use_system_prompt=True, reasoning_effort="low")
-    elif name == "gpt_oss_medium_reasoning":
-        return GptOssRenderer(tokenizer, use_system_prompt=True, reasoning_effort="medium")
-    elif name == "gpt_oss_high_reasoning":
-        return GptOssRenderer(tokenizer, use_system_prompt=True, reasoning_effort="high")
-    else:
-        raise ValueError(f"Unknown renderer: {name}")
